hallucination-detector/requirements.txt
google-genai>=1.0.0
faiss-cpu>=1.8.0
fastapi>=0.111.0
uvicorn>=0.29.0
pypdf>=4.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
